Re-organized the module/package structure
After discovering a bug in DmapRead that only
DmapWrite could detect because of the SuperDARN
file type checking. This made it clear that a
SuperDARN read methods needs to be created thus a
new module was needed. Separated modules and create
DarnRead and DarnWrite for SuperDARN file types. These classes
inherit from DmapRead and DmapWrite for the underlying
file type format. Exceptions were also separated for
better modularity.
