+++ pydarn/__init__.py
@@ -13,17 +13,18 @@ import logging.config
 import yaml
 
 # Importing pydarn exception classes
-from .exceptions import pydmap_exceptions
+from .exceptions import dmap_exceptions
+from .exceptions import superdarn_exceptions
 
 # Importing pydarn pydmap data structure classes
-from .pydmap.datastructures import DmapScalar
-from .pydmap.datastructures import DmapArray
-from .pydmap.datastructures import DmapRecord
-from .pydmap import superdarn_formats
+from .io.datastructures import DmapScalar
+from .io.datastructures import DmapArray
+from .io.datastructures import DmapRecord
+from .io import superdarn_formats
 
 # Importing pydarn pydmap classes
-from .pydmap.io import DmapRead
-from .pydmap.io import DmapWrite
+from .io.dmap import DmapRead
+from .io.dmap import DmapWrite
 
 """
 Pydarn uses yaml for logging configuration because it is the
